Move enums to Enums folder
