+++ Files/Helpers/Convert/ErrorCodeConverter.cs
@@ -1,4 +1,5 @@
-﻿using Files.Filesystem;
+﻿using Files.Enums;
+using Files.Filesystem;
 
 namespace Files.Helpers
 {
